% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_glucose.R
\name{approx_glucose}
\alias{approx_glucose}
\title{Approximate Glucose Data}
\usage{
approx_glucose(glucose, interval = 1, interval_unit = "mins",
  round_second = TRUE, ...)
}
\arguments{
\item{glucose}{A \code{data.frame} with a time column and glucose column}

\item{interval}{The interval to space out the times}

\item{interval_unit}{the unit for the intervals}

\item{round_second}{Round the seconds in the times}

\item{...}{additional arguments passed to \code{\link{approxfun}}}
}
\value{
A \code{data.frame} of all the times in the range of the original data
}
\description{
Creates a \code{data.frame} of all the times interpolated to
the interval.   Uses \code{\link{approxfun}} to approximate
}
