% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_time_df.R
\name{complete_time_df}
\alias{complete_time_df}
\title{Complete all time intervals for a recording}
\usage{
complete_time_df(df, interval = NULL, interval_unit = NULL)
}
\arguments{
\item{df}{A \code{data.frame} with a time column to be merged upon}

\item{interval}{The interval to space out the times}

\item{interval_unit}{the unit for the intervals}
}
\value{
A \code{data.frame} of all the times in a sequcne
}
\description{
Creates all the times for a certain interval and merges
with a current data set
}
\note{
If either \code{interval} or \code{interval_unit} are specified, the
other must be specified.  If both are \code{NULL}, then
\code{\link{interval_check}} is run and the minimum interval is used
}
