% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_intervals.R
\name{create_all_intervals}
\alias{create_all_intervals}
\alias{create_all_intervals_df}
\title{Create all time intervals for a recording}
\usage{
create_all_intervals(times, interval = 1, interval_unit = "mins")

create_all_intervals_df(...)
}
\arguments{
\item{times}{Vector of times, usually POSIXct or other date format}

\item{interval}{The interval to space out the times}

\item{interval_unit}{the unit for the intervals}

\item{...}{not used.}
}
\value{
A vector of times from minimum to the maximum
(including some potential roundup)
}
\description{
Creates all the times for a certain interval
}
\note{
The additional \code{create_all_intervals_df} function is
to simply put this vector into a \code{data.frame} in a column called
\code{time} and a logical called \code{grid}
}
