% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_numeric.R
\name{ensure_numeric}
\alias{ensure_numeric}
\title{Ensure a numeric output}
\usage{
ensure_numeric(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
vector of numeric values
}
\description{
Simple wrapper for character to numeric conversion
}
