% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_24.R
\name{first_24}
\alias{first_24}
\title{Flag DEXCOM CGMS first 24 hours}
\usage{
first_24(glucose)
}
\arguments{
\item{glucose}{\code{data.frame} of glucose data with a column "time"}
}
\value{
A \code{data.frame} of values with additional columns for within
first 24 hours and time from baseline
}
\description{
Flags points of data within the first 24 hours of wear.
}
