% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abbott.R
\name{read_abbott}
\alias{read_abbott}
\title{Read Abbott Output Data}
\usage{
read_abbott(path)
}
\arguments{
\item{path}{Path of .xlsx file.  Must have a sheet named "AbbottRaw" in it}
}
\value{
If the sheet is not empty, it will return a \code{data.frame} of
values.  Otherwise, it will return \code{NULL}
}
\description{
Reads an excel file with Abbot Raw data
}
