% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dexcom.R
\name{read_dexcom}
\alias{read_dexcom}
\title{Read Dexcom Output Data}
\usage{
read_dexcom(path, raw_sheet = TRUE, complete = TRUE, ...)
}
\arguments{
\item{path}{Path of .xlsx file.  Must have a sheet named "DexcomRaw" in it}

\item{raw_sheet}{Should the raw or processed sheet be read in?}

\item{complete}{Should all the times be completed for the data?}

\item{...}{additional arguments passed to \code{complete_time_df}}
}
\value{
If the sheet is not empty, it will return a \code{data.frame} of
values.  Otherwise, it will return \code{NULL}
}
\description{
Reads an excel file with Dexcom Raw data
}
