% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dexcom_export.R
\name{read_dexcom_export}
\alias{read_dexcom_export}
\title{Read Dexcom Export Data}
\usage{
read_dexcom_export(path, complete = FALSE, make_numeric = TRUE, ...)
}
\arguments{
\item{path}{Path of .txt file}

\item{complete}{Should all the times be completed for the data?}

\item{make_numeric}{Should glucose be coerced to numeric?  Values such as
"\code{Low}" would be missing in the glucose data}

\item{...}{additional arguments passed to \code{\link{complete_time_df}}}
}
\value{
If the sheet is not empty, it will return a \code{data.frame} of
values.  Otherwise, it will return \code{NULL}
}
\description{
Reads a raw Dexcom export
}
