% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gluc.R
\name{read_gluc}
\alias{read_gluc}
\title{Read Glucose XLS Data file}
\usage{
read_gluc(path, type = c("Abbott", "Dexcom", "AbbottRaw", "DexcomRaw"), ...)
}
\arguments{
\item{path}{path of xlsx file}

\item{type}{type of file to read}

\item{...}{additional argument to send to the read function related to
\code{type}}

\item{verbose}{print diagnostic messages}
}
\value{
A \code{tbl} data.frame type.
}
\description{
A global function to read the different types of glucose
data
}
