% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gluc_all.R
\name{read_gluc_all}
\alias{read_gluc_all}
\title{Read All Glucose Sheets from file}
\usage{
read_gluc_all(path, type = c("Abbott", "Dexcom", "AbbottRaw", "DexcomRaw"),
  ...)
}
\arguments{
\item{path}{paths of xlsx files}

\item{type}{type of file to read, defaults to all}

\item{...}{additional argument to send to the read function related to
\code{type}}
}
\value{
A list of \code{tbl} data.frame for each type.
}
\description{
A global function to read the different types of glucose
data from an xlsx
}
