% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_check.R
\name{sheet_check}
\alias{sheet_check}
\title{Check and Complete a Sheet}
\usage{
sheet_check(df, complete = FALSE, ...)
}
\arguments{
\item{df}{\code{data.frame} of the sheet with \code{time} as a column}

\item{complete}{Should all the times be completed for the data?}

\item{...}{additional arguments passed to \code{\link{complete_time_df}}}
}
\value{
Either the completed data.frame or NULL
}
\description{
Wrapper check for Abbott and Dexcom sheets
}
