% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abbott.R
\name{read_abbott}
\alias{read_abbott}
\title{Read Abbott Output Data}
\usage{
read_abbott(path, raw_sheet = TRUE, complete = FALSE, make_numeric = TRUE,
  ...)
}
\arguments{
\item{path}{Path of .xlsx file.  Must have a sheet named "AbbottRaw" in it}

\item{raw_sheet}{Should the raw or processed sheet be read in?}

\item{complete}{Should all the times be completed for the data?}

\item{make_numeric}{Should glucose be coerced to numeric?  Values such as
"\code{Low}" would be missing in the glucose data}

\item{...}{additional arguments passed to \code{\link{complete_time_df}}}
}
\value{
If the sheet is not empty, it will return a \code{data.frame} of
values.  Otherwise, it will return \code{NULL}
}
\description{
Reads an excel file with Abbot Raw data
}
